seeded

+++ app/models/asset.py
@@ -11,7 +11,7 @@ class Asset(db.Model):
     employees = db.Column(db.Integer, nullable=False)
     headquarters = db.Column(db.String, nullable=False)
     founded = db.Column(db.Integer, nullable=False)
-    market_cap = db.Column(db.Integer, nullable=False)
+    market_cap = db.Column(db.BigInteger, nullable=False)
     price_earning_ratio = db.Column(db.Integer, nullable=False)
     dividend_yield = db.Column(db.Integer, nullable=False)
     average_volume = db.Column(db.Integer, nullable=False)
